+++ src/modules/Plugins.js
@@ -7,7 +7,7 @@ Warden.Worker = function(adr){
   }
   stream.post = worker.postMessage;
   stream.onmessage = worker.onmessage
-  return stream;
+  return stream.newBust();
 }
 
 Warden.Observe = function(obj){
@@ -16,6 +16,7 @@ Warden.Observe = function(obj){
     Object.observe(obj, function(){
       stream.eval.apply(obj, arguments);
     })
+    return stream.newBus();
   }else{
     throw "This browser doesn't implement Object.observe"
   }